+++ 老師的程式碼/MyWebAPI/Controllers/ProductsController.cs
@@ -8,6 +8,7 @@ using Microsoft.Data.SqlClient;
 using Microsoft.EntityFrameworkCore;
 using MyWebAPI.DTOs;
 using MyWebAPI.Models;
+using MyWebAPI.Services;
 
 namespace MyWebAPI.Controllers
 {
@@ -26,16 +27,19 @@ namespace MyWebAPI.Controllers
 
         //4.7.8 修改ProductsController上方所注入的GoodStoreContext為GoodStoreContext2
         private readonly GoodStoreContextG2 _context;
+        private readonly SomeService _service;
 
-        public ProductsController(GoodStoreContextG2 context)
+        public ProductsController(GoodStoreContextG2 context, SomeService service)
         {
             _context = context;
+            _service = service;
         }
 
         // GET: api/Products
         [HttpGet()]
         public async Task<ActionResult<IEnumerable<ProductDTO>>> GetProduct(string? cateID, string? productName, decimal? minPrice, decimal? maxPrice, string? description)
         {
+
             //4.1.2 使用Include()同時取得關聯資料
             //4.1.3 使用Where()改變查詢的條件並測試
             //4.1.4 使用OrderBy()相關排序方法改變資料排序並測試
@@ -216,10 +220,10 @@ namespace MyWebAPI.Controllers
             return product;
         }
 
-        // PUT: api/Products/5
-        // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
+
+        //6.1.7 改寫ProductsController中Put Action內容
         [HttpPut("{id}")]
-        public async Task<IActionResult> PutProduct(string id, [FromForm]ProductPutDTO product)
+        public async Task<IActionResult> PutProduct(string id, [FromForm] ProductPutDTO product)
         {
             //if (id != product.ProductID)
             //{
@@ -237,7 +241,7 @@ namespace MyWebAPI.Controllers
             }
 
             //檢查是否有新照片上傳
-            if (product.Picture != null || product.Picture.Length != 0)
+            if (product.Picture != null && product.Picture.Length != 0)
             {
                 FileUpload(product.Picture, id);
 
@@ -306,7 +310,7 @@ namespace MyWebAPI.Controllers
 
             string fileName = await FileUpload(product.Picture, product.ProductID);
 
-            if(fileName=="")
+            if (fileName == "")
             {
                 return BadRequest("上傳的檔案格式不正確，請上傳jpg、jpeg或png格式的圖片");
             }
@@ -346,7 +350,7 @@ namespace MyWebAPI.Controllers
             return CreatedAtAction("GetProduct", new { id = product.ProductID }, product);
         }
 
-        // DELETE: api/Products/5
+        //7.1.1 改寫ProductsController中Delete Action內容，加入刪除照片的功能
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteProduct(string id)
         {
@@ -356,11 +360,55 @@ namespace MyWebAPI.Controllers
                 return NotFound();
             }
 
+            //刪除商品照片
+            if(! await FileDelete(product.Picture))
+            {
+                return BadRequest("刪除商品照片失敗，請檢查檔案是否存在或權限問題。");
+            }
+
+
             _context.Product.Remove(product);
             await _context.SaveChangesAsync();
 
             return NoContent();
         }
+
+        //7.1.5 建立可刪除多筆資料的Delete Action(批次刪除)，介接口設為[HttpDelete("ByCatID")]，方法名稱可自訂，傳入的參為為商品類別ID
+        //方法名稱可自訂，傳入的參為為商品類別ID
+        [HttpDelete("ByCateID")]
+        public async Task<IActionResult> DeleteProductsByCateID(string cateID)
+        {
+            var products = await _context.Product.Where(p => p.CateID == cateID).ToListAsync();
+
+            if (products == null)
+            {
+                return NotFound();
+            }
+
+
+            foreach(var p in products)
+            {
+                //刪除商品照片
+                if (!await FileDelete(p.Picture))
+                {
+                    return BadRequest("刪除商品照片失敗，請檢查檔案是否存在或權限問題。");
+                }
+                //刪除商品資料
+                _context.Product.Remove(p);
+            }
+
+            try
+            {
+                await _context.SaveChangesAsync();
+            }
+            catch (DbUpdateException)
+            {
+                return NotFound("刪除商品失敗，請檢查商品是否存在。");
+            }
+
+            return NoContent();
+        }
+
 
         private bool ProductExists(string id)
         {
@@ -423,6 +471,37 @@ namespace MyWebAPI.Controllers
 
 
             return fileName; //回傳檔案名稱
+        }
+
+
+        //7.1.2 將刪除照片功能另建立FileDelete()方法
+        private async Task<bool> FileDelete(string fileName)
+        {
+
+            var path = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot", "ProductPhotos");
+
+            var filePath = Path.Combine(path, fileName);
+
+            if(System.IO.File.Exists(filePath))
+            {
+                try
+                {
+                    System.IO.File.Delete(filePath);
+
+                    return true;
+                }
+                catch (Exception ex)
+                {
+                  
+                  
+                    return false;
+                }
+            }
+
+
+          
+
+            return false;
         }
     }
 }
